PLB2B-891: Rename framework scheme
Experiments have shown that Carthage prepares wrong project setup and starts building scheme as part of project instead of workspace.
This is issue was gone after renaming the scheme. There are 2 targets called SecureAccessBLE, one in Pods project and one in SecureAccessBLE project,
I assume something gets confused if the scheme is named the same. Should be a bug.